requirements.txt
ccxt>=4.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pyyaml>=6.0
ta-lib>=0.4.28
python-dotenv>=1.0.0
loguru>=0.7.0
matplotlib>=3.7.0
mplfinance>=0.12.10
pytz>=2023.3
pyarrow>=10.0.0,<22.0.0

